+++ pallets/risk-manager/src/mock.rs
@@ -21,7 +21,6 @@ use std::thread;
 pub use test_helper::*;
 
 pub const PROTOCOL_INTEREST_TRANSFER_THRESHOLD: Balance = 1_000_000_000_000_000_000_000;
-pub type AccountId = u64;
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
 
@@ -160,17 +159,9 @@ impl Contains<u64> for WhitelistMembers {
 }
 
 pub const ONE_HUNDRED: Balance = 100;
-pub const DOLLARS: Balance = 1_000_000_000_000_000_000;
-pub fn dollars(amount: u128) -> u128 {
-	amount.saturating_mul(Price::accuracy())
-}
 pub const ADMIN: AccountId = 0;
 pub fn admin() -> Origin {
 	Origin::signed(ADMIN)
-}
-pub const ALICE: AccountId = 1;
-pub fn alice() -> Origin {
-	Origin::signed(ALICE)
 }
 pub const BOB: AccountId = 2;
 pub const CHARLIE: AccountId = 3;
